Change eta to double, e.g. arte provides the ETA as double. Round double durations to avoid having too many decimal places in the UI

+++ src/main/kotlin/de/lobbenmeier/stefan/ytdlp/DownloadProgress.kt
@@ -42,7 +42,7 @@ data class YtDlpDownloadProgress(
     val totalBytesEstimate: Float? = null,
     val tmpfilename: String? = null,
     val filename: String? = null,
-    val eta: Int? = null,
+    val eta: Double? = null,
     val speed: Float? = null,
     val elapsed: Float? = null,
     val ctxId: String? = null,
